Command Line Update
Added Command line handling so that the code doesn't have to be changed to download a different feed.

Also make the number of downloads a command line parameter, default is all.

Directory to download to is also a parameter now.

Added --Title to optionally use the episode title as the file name,

Update the documentation